# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "twp server api.",
        "title": "twp API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.o"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/buyer/cart/{cart_id}/coupon": {
            "get": {
                "description": "Buyer get usable coupon of cart/shop",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Buyer",
                    "Cart",
                    "Coupon"
                ],
                "summary": "Buyer Get usable coupon of cart/shop",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Cart ID",
                        "name": "cart_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/db.GetUsableCouponsRow"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
        "/buyer/cart/{cart_id}/coupon/{coupon_id}": {
            "post": {
                "description": "Add coupon to cart",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Buyer",
                    "Cart",
                    "Coupon"
                ],
                "summary": "Buyer Add Coupon To Cart",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Cart ID",
                        "name": "cart_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Coupon ID",
                        "name": "coupon_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete coupon In cart",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Buyer",
                    "Cart",
                    "Coupon"
                ],
                "summary": "Buyer Delete Coupon In Cart",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Cart ID",
                        "name": "cart_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Coupon ID",
                        "name": "coupon_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
...
        "/buyer/order": {
            "get": {
                "description": "Get all order history of the user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Buyer",
                    "Order"
                ],
                "summary": "Buyer Get Order History",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Begin index",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/db.GetOrderHistoryRow"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
        "/buyer/order/{id}": {
            "get": {
                "description": "Get specific order detail",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Buyer",
                    "Order"
                ],
                "summary": "Buyer Get Order Detail",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/buyer.OrderDetail"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "buyer.OrderDetail": {
            "type": "object",
            "properties": {
                "details": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.GetOrderDetailRow"
                    }
                },
                "info": {
                    "$ref": "#/definitions/db.GetOrderInfoRow"
                }
            }
        },
...
        "db.CouponScope": {
            "type": "string",
            "enum": [
                "global",
                "shop"
            ],
            "x-enum-varnames": [
                "CouponScopeGlobal",
                "CouponScopeShop"
            ]
        },
...
        "db.CouponType": {
            "type": "string",
            "enum": [
                "percentage",
                "fixed",
                "shipping"
            ],
            "x-enum-varnames": [
                "CouponTypePercentage",
                "CouponTypeFixed",
                "CouponTypeShipping"
            ]
        },
...
        "db.GetOrderDetailRow": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "image_url": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "product_id": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
        "db.GetOrderHistoryRow": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "shipment": {
                    "type": "integer"
                },
                "shop_image_url": {
                    "type": "string"
                },
                "shop_name": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/db.OrderStatus"
                },
                "thumbnail_url": {
                    "type": "string"
                },
                "total_price": {
                    "type": "integer"
                }
            }
        },
        "db.GetOrderInfoRow": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "discount": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "shipment": {
                    "type": "integer"
                },
                "shop_image_url": {
                    "type": "string"
                },
                "shop_name": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/db.OrderStatus"
                },
                "total_price": {
                    "type": "integer"
                }
            }
        },
...
        "db.GetUsableCouponsRow": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "discount": {
                    "type": "number"
                },
                "expire_date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "scope": {
                    "$ref": "#/definitions/db.CouponScope"
                },
                "type": {
                    "$ref": "#/definitions/db.CouponType"
                }
            }
        },
...
        "db.OrderStatus": {
            "type": "string",
            "enum": [
                "paid",
                "shipped",
                "delivered",
                "cancelled",
                "finished"
            ],
            "x-enum-varnames": [
                "OrderStatusPaid",
                "OrderStatusShipped",
                "OrderStatusDelivered",
                "OrderStatusCancelled",
                "OrderStatusFinished"
            ]
        },
...
        "echo.HTTPError": {
            "type": "object",
            "properties": {
                "message": {}
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    },
...
}